+++ packages/agent_dart_base/lib/agent/crypto/keystore/key_store.dart
@@ -6,10 +6,10 @@ import 'dart:typed_data';
 
 import 'package:agent_dart_base/agent/cbor.dart';
 import 'package:agent_dart_base/agent/crypto/random.dart';
-import 'package:agent_dart_base/src/ffi/io.dart' hide SHA256;
 import 'package:agent_dart_base/identity/p256.dart';
 import 'package:agent_dart_base/identity/secp256k1.dart';
 import 'package:agent_dart_base/principal/utils/sha256.dart';
+import 'package:agent_dart_base/src/ffi/io.dart';
 import 'package:agent_dart_base/utils/extension.dart';
 import 'package:agent_dart_base/utils/u8a.dart';
 import 'package:agent_dart_ffi/agent_dart_ffi.dart';
@@ -18,10 +18,8 @@ import 'package:meta/meta.dart';
 import 'package:pointycastle/export.dart';
 import 'package:uuid/uuid.dart';
 
-part 'key_derivator.dart';
-
 part 'function.dart';
-
+part 'key_derivator.dart';
 part 'util.dart';
 
 @immutable