+++ packages/agent_dart_base/test/agent_dart_base_test.dart
@@ -1,5 +1,3 @@
-import 'package:agent_dart_base/src/ffi/io.dart';
-
 import 'agent/agent.dart' as agent;
 import 'authentication/authentication.dart' as auth;
 import 'candid/idl.dart' as candid;
@@ -9,8 +7,6 @@ import 'test_utils.dart';
 import 'utils/utils_test.dart' as utils;
 import 'wallet/pem.dart' as pem;
 import 'wallet/signer.dart' as signer;
-import 'dart:io' as io;
-import 'dart:ffi';
 
 void main() async {
   matchFFI();